+++ core/src/main/scala/neotypes/Async.scala
@@ -1,8 +1,5 @@
 package neotypes
 
-import scala.concurrent.{ExecutionContext, Future, Promise}
-import scala.util.{Failure, Success}
-
 trait Async[F[_]] {
   def async[T](cb: (Either[Throwable, T] => Unit) => Unit): F[T]
 
@@ -15,44 +12,4 @@ trait Async[F[_]] {
   def failed[T](e: Throwable): F[T]
 
   def success[T](t: => T): F[T]
-}
-
-object Async {
-  implicit class AsyncExt[F[_], T](val m: F[T]) extends AnyVal {
-    def map[U](f: T => U)(implicit F: Async[F]): F[U] =
-      F.map(m)(f)
-
-    def flatMap[U](f: T => F[U])(implicit F: Async[F]): F[U] =
-      F.flatMap(m)(f)
-
-    def recoverWith[U >: T](f: PartialFunction[Throwable, F[U]])(implicit F: Async[F]): F[U] =
-      F.recoverWith[T, U](m)(f)
-  }
-
-  implicit def futureAsync(implicit ec: ExecutionContext): Async[Future] =
-    new Async[Future] {
-      override def async[A](cb: (Either[Throwable, A] => Unit) => Unit): Future[A] = {
-        val p = Promise[A]()
-        cb {
-          case Right(res) => p.complete(Success(res))
-          case Left(ex)   => p.complete(Failure(ex))
-        }
-        p.future
-      }
-
-      override def flatMap[T, U](m: Future[T])(f: T => Future[U]): Future[U] =
-        m.flatMap(f)
-
-      override def map[T, U](m: Future[T])(f: T => U): Future[U] =
-        m.map(f)
-
-      override def recoverWith[T, U >: T](m: Future[T])(f: PartialFunction[Throwable, Future[U]]): Future[U] =
-        m.recoverWith(f)
-
-      override def failed[T](e: Throwable): Future[T] =
-        Future.failed(e)
-
-      override def success[T](t: => T): Future[T] =
-        Future.successful(t)
-    }
 }
